python/requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.3.0
async-timeout==4.0.3
attrs==24.2.0
blinker==1.8.2
certifi==2024.7.4
click==8.1.7
dataclass-factory==2.16
distro==1.9.0
exceptiongroup==1.2.2
ExifRead==3.0.0
Flask==3.0.3
Flask-Cors==4.0.2
frozenlist==1.4.1
geographiclib==2.0
geopy==2.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib_metadata==8.2.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
MarkupSafe==2.1.5
multidict==6.1.0
numpy==1.26.4
openai==1.40.0
opencv-python==4.10.0.84
packaging==24.1
pillow==10.4.0
pydantic==1.10.18
pydantic_core==2.20.1
pydub==0.25.1
pytesseract==0.3.10
shazamio==0.6.0
shazamio_core==1.0.7
sniffio==1.3.1
tesseract==0.1.3
tqdm==4.66.5
typing_extensions==4.12.2
Werkzeug==3.0.3
yarl==1.13.1
zipp==3.20.0
